Only try to remove invalid filter from network if it's our own. This avoids unnecessary/confusing warning logs in data storage.

+++ core/src/main/java/bisq/core/filter/FilterManager.java
@@ -314,13 +314,18 @@ public class FilterManager {
     }
 
     public void removeInvalidFilters(Filter filter, String privKeyString) {
-        log.info("Remove invalid filter {}", filter);
-        setFilterSigningKey(privKeyString);
-        String signatureAsBase64 = getSignature(Filter.cloneWithoutSig(filter));
-        Filter filterWithSig = Filter.cloneWithSig(filter, signatureAsBase64);
-        boolean result = p2PService.removeData(filterWithSig);
-        if (!result) {
-            log.warn("Could not remove filter {}", filter);
+        // We can only remove the filter if it's our own filter
+        if (Arrays.equals(filter.getOwnerPubKey().getEncoded(), keyRing.getSignatureKeyPair().getPublic().getEncoded())) {
+            log.info("Remove invalid filter {}", filter);
+            setFilterSigningKey(privKeyString);
+            String signatureAsBase64 = getSignature(Filter.cloneWithoutSig(filter));
+            Filter filterWithSig = Filter.cloneWithSig(filter, signatureAsBase64);
+            boolean result = p2PService.removeData(filterWithSig);
+            if (!result) {
+                log.warn("Could not remove filter {}", filter);
+            }
+        } else {
+            log.info("The invalid filter is not our own, so we cannot remove it from the network");
         }
     }
 
@@ -523,13 +528,13 @@ public class FilterManager {
 
         if (currentFilter != null) {
             if (currentFilter.getCreationDate() > newFilter.getCreationDate()) {
-                log.debug("We received a new filter from the network but the creation date is older than the " +
+                log.info("We received a new filter from the network but the creation date is older than the " +
                         "filter we have already. We ignore the new filter.");
 
                 addToInvalidFilters(newFilter);
                 return;
             } else {
-                log.debug("We received a new filter from the network and the creation date is newer than the " +
+                log.info("We received a new filter from the network and the creation date is newer than the " +
                         "filter we have already. We ignore the old filter.");
                 addToInvalidFilters(currentFilter);
             }
@@ -580,7 +585,7 @@ public class FilterManager {
 
         // We don't check for banned filter as we want to remove a banned filter anyway.
 
-        if (!filterProperty.get().equals(filter)) {
+        if (filterProperty.get() != null && !filterProperty.get().equals(filter)) {
             return;
         }
